requirements.txt
google-genai==1.22.0
python-dotenv==1.1.1
streamlit==1.46.1

